Removing lucrasoft.nl email addresses from code. - the "no-reply@lucrasoft.nl" email address was replaced with a configuration setting
- Configuration settings redirectMails/redirectTo were removed.  This config settings were made to allow testing email with a real mailserver without mailing a real user.
That approach had the downside of hard-coding a configuration useful only to past developers.
A better approach is for localhost to try and use localhost as an email server.  By doing that, integration tests can run a fake server to test behavior.  Developers can run a localhost mailserver for manual testing (fake or not, something like http://smtp4dev.codeplex.com/ is best).